# Api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Voting service",
        "title": "Voting service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/office/create-office": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new office based on the provided office details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Office"
                ],
                "summary": "Create a new office",
                "parameters": [
                    {
                        "description": "Create office request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.CreateOfficeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Office created successfully",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    }
                }
            }
        },
        "/office/delete-office": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an office based on the provided office ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Office"
                ],
                "summary": "Delete an office",
                "parameters": [
                    {
                        "description": "Delete office request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.DeleteOfficeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Office deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    },
                    "404": {
                        "description": "Office not found",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    }
                }
            }
        },
        "/office/get-all-ofice": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all offices with optional pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Office"
                ],
                "summary": "Get all offices",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all offices retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/delivery.GetAllOfficesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    }
                }
            }
        },
        "/office/get-office": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves details of an office based on the provided office ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Office"
                ],
                "summary": "Get office details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Office ID",
                        "name": "office_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Office details retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/delivery.OfficeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    }
                }
            }
        },
        "/office/update-office": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the details of an existing office based on the provided office ID and details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Office"
                ],
                "summary": "Update an existing office",
                "parameters": [
                    {
                        "description": "Update office request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.UpdateOfficeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Office updated successfully",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    },
                    "404": {
                        "description": "Office not found",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.InfoResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "delivery.CreateOfficeRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "delivery.DeleteOfficeRequest": {
            "type": "object",
            "properties": {
                "office_id": {
                    "type": "string"
                }
            }
        },
...
        "delivery.GetAllOfficesResponse": {
            "type": "object",
            "properties": {
                "offices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/delivery.OfficeResponse"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "delivery.InfoResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "delivery.OfficeResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "office_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "delivery.UpdateOfficeRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "office_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}